# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://task.firscoders.shop/api/",
            "description": "Global API server"
        },
        {
            "url": "http://test-task.loc/api/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/v1/orders": {
            "get": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Orders list",
                "description": "Orders list.",
                "operationId": "ordersList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "15"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Order add",
                "description": "Order add.",
                "operationId": "OrderCreate",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "43"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/orders/{order_id}": {
            "get": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Order show",
                "description": "Order show.",
                "operationId": "orderShow",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Order update",
                "description": "Order update.",
                "operationId": "orderUpdate",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "4"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "43"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Order delete",
                "description": "Order delete.",
                "operationId": "orderDelete",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/orders/{order_id}/edit": {
            "get": {
                "tags": [
                    "Order Routes"
                ],
                "summary": "Order edit info",
                "description": "Order edit info.",
                "operationId": "orderEdit",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}